advance-agent/pyproject.toml
[project]
name = "advance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "firecrawl>=4.18.0",
    "firecrawl-py>=4.18.0",
    "langchain>=1.2.10",
    "langchain-core>=1.2.17",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.10",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]
